+++ server/src/main/java/org/opensearch/search/aggregations/metrics/MinAggregator.java
@@ -58,8 +58,6 @@ import org.opensearch.search.aggregations.support.ValuesSource;
 import org.opensearch.search.aggregations.support.ValuesSourceConfig;
 import org.opensearch.search.internal.SearchContext;
 import org.opensearch.search.startree.StarTreeQueryHelper;
-import org.opensearch.search.streaming.Streamable;
-import org.opensearch.search.streaming.StreamingCostMetrics;
 
 import java.io.IOException;
 import java.util.Map;
@@ -73,7 +71,7 @@ import static org.opensearch.search.startree.StarTreeQueryHelper.getSupportedSta
  *
  * @opensearch.internal
  */
-class MinAggregator extends NumericMetricsAggregator.SingleValue implements StarTreePreComputeCollector, Streamable {
+class MinAggregator extends NumericMetricsAggregator.SingleValue implements StarTreePreComputeCollector {
     private static final int MAX_BKD_LOOKUPS = 1024;
 
     final ValuesSource.Numeric valuesSource;
@@ -305,10 +303,5 @@ class MinAggregator extends NumericMetricsAggregator.SingleValue implements Star
     @Override
     public void doReset() {
         mins.fill(0, mins.size(), Double.POSITIVE_INFINITY);
-    }
-
-    @Override
-    public StreamingCostMetrics getStreamingCostMetrics() {
-        return new StreamingCostMetrics(true, 1, 1, 1, 1);
     }
 }