+++ test/framework/src/main/java/org/opensearch/search/aggregations/AggregatorTestCase.java
@@ -149,6 +149,7 @@ import org.opensearch.search.internal.ContextIndexSearcher;
 import org.opensearch.search.internal.SearchContext;
 import org.opensearch.search.lookup.SearchLookup;
 import org.opensearch.search.startree.StarTreeQueryContext;
+import org.opensearch.search.streaming.FlushMode;
 import org.opensearch.test.InternalAggregationTestCase;
 import org.opensearch.test.OpenSearchTestCase;
 import org.junit.After;
@@ -332,6 +333,9 @@ public abstract class AggregatorTestCase extends OpenSearchTestCase {
     ) throws IOException {
         SearchContext searchContext = createSearchContext(indexSearcher, indexSettings, query, bucketConsumer, fieldTypes);
         when(searchContext.isStreamSearch()).thenReturn(true);
+        // Force streaming aggregator creation by setting flushMode to PER_SEGMENT
+        // This bypasses the cost estimation decision logic in the factory
+        when(searchContext.getFlushMode()).thenReturn(FlushMode.PER_SEGMENT);
         return createAggregator(aggregationBuilder, searchContext);
     }